Fix a code-edit whitespace case
Fixes #12567.
