+++ app/ydoc-shared/src/ast/documentation.ts
@@ -19,7 +19,7 @@ export function* docLineToConcrete(
   let extraIndent = ''
   for (const { token } of docLine.docs.elements) {
     if (token.node.tokenType_ === TokenType.Newline) {
-      yield ensureUnspaced(token, false)
+      yield preferUnspaced(token)
     } else {
       if (prevType === TokenType.Newline) {
         yield { whitespace: token.whitespace ?? indent + extraIndent, node: token.node }
